requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
PyMuPDF==1.27.2
pdfplumber==0.11.4
Pillow==11.0.0
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
openai==1.51.0
aiofiles==24.1.0
